change the test file name and update the test cases
